refactor: STEP3 리뷰 반영(패키지 이동)
