+++ src/main/java/lotto/model/LottoPapers.java
@@ -1,7 +1,5 @@
 package lotto.model;
 
-import lotto.controller.LottoNumberGenerator;
-
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
