+++ src/test/java/lotto/model/LottoMoneyTest.java
@@ -1,4 +1,4 @@
-package lotto.controller;
+package lotto.model;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
